Fix a few issues in sortCars. It's starting to look pretty stable!

+++ main.h
@@ -144,10 +144,10 @@ typedef struct thisShouldBeAnObjectOnSoManyLevels
 	
 	struct
 	{
-		byte currentLine;
-		byte currentColumn;
+		uint currentLine;
+		uint currentColumn;
 		
-		byte currentIndex;
+		uint currentIndex;
 		
 		byte section;
